sql types in separate header files
